Add the ability for react-props to keep existing propTypes
When migrating a codebase to TypeScript in one go, it makes sense to
delete PropType declarations in favor of TypeScript types. However, in
the case of partial migrations, TypeScript types might not be sufficient
alone. A migrated TSX file might still be used in a JS file, which would
make the benefits of its types lost (unless --checkJs is specified).

This commit adds the ability to keep propTypes when running the
react-props plugin. This option would allow a migrated file to retain
its PropTypes validations, while still benefitting from TypeScript's
types. It is disabled by default, and can be enabled with the
`shouldKeepPropTypes` plugin option.

+++ packages/ts-migrate-plugins/src/plugins/react-props.ts
@@ -18,6 +18,7 @@ type Options = {
   anyAlias?: string;
   anyFunctionAlias?: string;
   shouldUpdateAirbnbImports?: boolean;
+  shouldKeepPropTypes?: boolean;
 };
 
 export type PropTypesIdentifierMap = { [property: string]: string };
@@ -63,17 +64,19 @@ const reactPropsPlugin: Plugin<Options> = {
       updatedSourceText,
       sourceFile.languageVersion,
     );
-    const importUpdates = updateImports(
-      updatedSourceFile,
-      spreadReplacements.map((cur) => cur.typeImport),
-      [
-        { moduleSpecifier: 'prop-types' },
-        ...(options.shouldUpdateAirbnbImports ? importReplacements : []),
-        ...(options.shouldUpdateAirbnbImports
-          ? spreadReplacements.map((cur) => cur.spreadImport)
-          : []),
-      ],
-    );
+    const importUpdates = !options.shouldKeepPropTypes
+      ? updateImports(
+          updatedSourceFile,
+          spreadReplacements.map((cur) => cur.typeImport),
+          [
+            { moduleSpecifier: 'prop-types' },
+            ...(options.shouldUpdateAirbnbImports ? importReplacements : []),
+            ...(options.shouldUpdateAirbnbImports
+              ? spreadReplacements.map((cur) => cur.spreadImport)
+              : []),
+          ],
+        )
+      : [];
     return updateSourceText(updatedSourceText, importUpdates);
   },
 };
@@ -226,7 +229,10 @@ function updatePropTypes(
             text: updateText,
           });
         }
-        updates.push(...deleteSfcPropTypes(node, sourceFile));
+
+        if (!options.shouldKeepPropTypes) {
+          updates.push(...deleteSfcPropTypes(node, sourceFile));
+        }
       }
     }
   } else {
@@ -259,7 +265,9 @@ function updatePropTypes(
           )}`,
         });
 
-        updates.push(...deleteClassPropTypes(node, sourceFile));
+        if (!options.shouldKeepPropTypes) {
+          updates.push(...deleteClassPropTypes(node, sourceFile));
+        }
       }
     }
   }
@@ -300,7 +308,7 @@ function updateObjectLiteral(
   propsTypeAlias = ts.moveSyntheticComments(propsTypeAlias, propsTypeNode);
 
   const varStatement = getParentVariableStatement(objectLiteral, sourceFile);
-  if (varStatement) {
+  if (varStatement && !options.shouldKeepPropTypes) {
     updates.push({
       kind: 'replace',
       index: varStatement.pos,
